backend/requirements.txt
Django==5.2.5
mysqlclient>=2.2.7
# Para Python 3.13 necesitas un wheel moderno de numpy (evitar compilación local):
numpy>=2.1.3
# OpenCV con wheels recientes (compatible con Python 3.13)
opencv-python>=4.10.0.84
python-dotenv>=1.0.1

djangorestframework
djangorestframework-simplejwt
# Opcional (may fallar en Windows/Python 3.13):
# face-recognition==1.3.0
# dlib (requerido por face-recognition) no siempre tiene wheel para 3.13.
# Si lo necesitas, instala un wheel precompilado compatible con tu versión de Python/Windows.
